+++ lib/view.dart
@@ -37,7 +37,7 @@ export 'package:flutter/foundation.dart' show kIsWeb, mustCallSuper, protected;
 
 /// MVC
 export 'package:mvc_pattern/mvc_pattern.dart'
-    show AppStatefulWidgetMVC, InheritedStateMVC, SetState;
+    show AppStatefulWidgetMVC, InheritedStateMixin, InheritedStateMVC, SetState;
 
 /// App
 export 'package:mvc_application/src/view/app.dart';